requirements.txt
Flask==2.3.3
cohere==4.37
PyPDF2==3.0.1
Werkzeug==2.3.7
reportlab==4.0.4
python-dotenv==1.0.0
mysql-connector-python==8.2.0